Extract non-essential methods from `PartialFace`
